requirements.txt
streamlit==1.24.0
azure-core==1.31.0
azure-ai-formrecognizer==3.2.0
openai==0.28.0
pyodbc
python-dotenv==1.0.0
Pillow==9.5.0
PyPDF2==3.0.1
aiohttp
async-timeout==4.0.2